Correctly solve the dialog window leak on screen rotation issue
If the user rotates the device, by default the activity is
destroyed and created again. If a dialog is showing and is not
dismissed, a window leak will occur. Originally this was solved
by setting android:configChanges="orientation", which prevented
the activity from being destroyed on rotations, but this causes
issues with switching layouts on device rotations.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>