Implemented BootSetup, RaceAlarm and Util classes
The RaceAlarm class handles reminding the user of the next race
if the user has reminders enabled. The BootSetup class will handle
any tasks that need to occur on boot complete, which is currently
just setting the alarm for the next race.  The Util class will be
used by the common functions used across multiple classes, which
is currently logging if debugging is enabled.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>