Only show Vibrate option if device supports it
Some products (like the Motorola Xoom) do not have a vibrator. On
those products, the Vibrate option should be hidden to avoid users
believing there is a bug in the application when the device doesn't
vibrate on reminders like they requested.

The hasVibrator method is only available on SDK 11 or later.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>