Do not allow application backups
By default, android:allowBackup is set to true. Need to disable
so users cannot use adb backup/restore to pull the files off,
make changes, and push them back up.

Everything about the user is restored when the connect to the
server, so no need to allow the user to do it manually.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>