Activity stack fixes
The activity stack for Pick The Winner was broken, and the problem
was made worse with the introduction of the widget. It was simple
for a user to end up with multiple instances of the app running.

This patch fixes the activity stack so the user actually gets what
they expect when clicking on the app icon or the widget and what
they expect when touching back or home.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>