Only use schedule from server. Closes #108
Previously, when a user installed the app, there was a preloaded
schedule. It could be (as likely was) outdated, but the user
wouldn't get the updated schedule until the next GCM push. They
could not (and shouldn't have to) manually update the schedule.

To avoid this problem completely, the app now retrieves the
schedule from the server if one is not present, and no longer
includes a snapshot of the schedule in the apk.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>