Move Play Store disconnect workaround to IabHelper. Closes #112
The workaround needs to be in IabHelper so flagEndAsync() can be
called. Previously, the Pick The Winner activity would catch the
exception and hide the button. This worked most of the time, but
in the case of issue #112, the service re-connected. Because of
this, since the async flag was still set, the IabHelper threw the
IllegalStateException.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>