Protect against scenario where null account name was used
Google Analytics reported a crash where a null account name was
passed to GAE.connect(). After some experimenting, a scenario
was discovered that saw AccountsActivity be destroyed before
onActivityResult() was called. To protect against this,
onSaveInstanceState() is used to save off mAccountName. To
ensure the app doesn't crash in case this protection fails,
a null check of mAccountName was added before passing it to
GAE.connect().

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>