Questions: Do not immediately advance to next race. Phase 1 of #97
Previously, as soon as the race started, the Questions tab advanced
to the next race. This was inconvenient for users who want to go
back and check what they submitted while the race was in progress.

Now, the Questions tab will continue to show the "recent" race,
then advance to the next race.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>