Remove hard-coded driver lists
The dynamic driver feature has been working great. During the
off-season, the server was updated to never return a null drivers
list, even if the admin fails to submit questions. Therefore, it
is no longer required to fall back on an old, outdated driver list.

This will allow the server to begin to reuse some old numbers
sometime in the future if necessary.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>