Save current widget race in shared prefs. Resolves #126
If the application is killed by Android, the static sRace variable
was lost. This becomes a problem if that race was "recent", because
on the next widget update, the widget would advance to the next
race too early. The race ID is now saved in a shared preference so
it can always be restored to the same state.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>