Copy PTW backend code into this repo
This was made possible due to the switch to Gradle and Android
Studio. Having all PTW code in a single repo allows additional
features such as Cloud Endpoints.

Code was copied instead of merged with a full git history because
the full git history mentioned GitHub issues in the old server
repo, which when imported into this repo would refer to incorrect
issues in this repo. See this for more info:
http://stackoverflow.com/questions/28139247/issue-tracker-management-when-merging-two-github-repos

From PickTheWinnerServer repo SHA1:
6ead37c1815cc4cea0212538d76da12cb2e20600

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>